Add batch sending and small thread sleep time
Previously, the receive thread would busy wait checking the socket
for new data.  This kept the queue from filling up, but utilized a
large amount of CPU time without accomplishing much.  When any
sleep delay was introduced the queue would fill up with drawings,
but they would still only be sent one at a time.
Now, the receive thread attempt to keep the queue empty and send
drawings together as a single message.  This dramatically reduces
the CPU use and does not have much impact on the responsiveness
from the previous version.