Rewrite conn for more encapsulation + enhancements
It now has a slightly different interface where it expects encoded
data, then drops a header on the front of it that it removes at the
other end before passing the encoded data back to the controller.

The Controller now always initialized a connection and attempts to
utilize it even in offline mode, but its methods are all no-ops if it
has not been initialized.