Add a text tool to the client
The encoding and decoding had to change slightly to accomodate the
text field, and there is a slightly different path in the event
handling logic for text entry since a pop-up box is needed to enter
the text.

Also, removed some excess code from the Drawing class that was sitting
around from an older version.