Substantial rework to the GUI structure
The GUI was changed to match my current preference for structuring
controllers and views.  The root window is a separate entity and
different screens in the application are views that swap in and out.

Additionally, the controller is passed to the main window and views.
This increases their logical coupling and the amount of work necessary
when refactoring one or both, but substantially decreases the amount
of boilerplate code to have an API for callback setting and function
generation by the controller for passing to the view.