Update connection to use an incoming queue
Rather than be given a callback, the connection now has a queue for
data it receives.  The new expectation is that the owner of the
connection will have a thread that will loop and process data from
this receive queue.  I still might encapsulate this behavior in the
connection itself, but I prefer the separation so that different
approaches can be taken to processing the incoming data.