Move towards defaulting to paint view
Changes are still ongoing, networking does not work right now as the
architecture shuffles around to facilitate doing away with the setup
screen.  The final version should have host and connect options
presented in a menu, so you can draw right away and there is no
longer an "offline" mode.  Maybe even save/load functionality down the
road.

Currently, I am updating to facilitate the new application state design
from chadlib.  This pulls data from the controller and gets passed to
all of the views, making it easier to read\modify a single, canonical
location of the data from anywhere in the application.

The next step is giving it ownership of the connection queues so that
information is more easily accessed from around the application and
will facilitate easier access and changes later, like a separate
draw queue after the receive queue.