Fix encoding/decoding of multiple drawings
Previously, drawings never queued up enough so that multiple drawings
were sent at once.  With the new connect system, drawings were
queuing up before a connection was made and then revealing the error.

The problem was that multiple drawings in a row were being treated as
text at the end of the first drawing, causing an unpack error, then
the decoded length would be 0, and the looped decoding would never
end, stuck on the same error over and over.

I refactored decoding so that the error breaks out of the loop, which
means future errors like this will at least not cause an infinite
loop.  Then I fixed how text is encoded and decoded so that the length
of the text is always included, rather than previously just triggering
text decoding if there was any data left after the standard drawing
data.  I think the cause was that I never updated the decoding when I
added batch sending, and incorrectly believed it was working though
that logic was never triggering.

A pair of tests have been added to detect future problems in batch
encoding/decoding.