Add undo tool
This is a decently large change, adjusting how events are handled,
adding a new drawing type, a new drawing method, and how drawing ids
are stored.

Instead of hanging on to only the last drawing id for draggable
drawings, now all ids are stored on a stack, and that stack is popped
from to perform undos.  This removed the need to have a special
mechanism to clear out the intermediate drawings.  Now the program
puts undos to be "drawn" after the intermediate drawings, starting
with the second motion event.

A great side effect of this change is that now the program sends all
of the intermediate drawings along with the undos, and connected users
can see all of the other user's actions while they drag out different
shapes.