Utilize the new chadlib controllers, add syncing
Update the design based on controllers with specific logic.  This
pulls a bit of generic connection logic out of the pypaint code.

Also, add a new sync "drawing" that is utilized to request drawings
from the connected user.  This allows users to get offline drawings
from others.

Known bugs are:
-one user loading a drawing and another attempting to sync and
    retrieve it does not work.  Only non-loaded drawings sync.
-syncing a large number of drawings results in unpacking errors most
    of the time, and any drawings after the error are not retrieved.