Fix drawing history management, update PaintCanvas
There was an error where loaded drawings were not being synced to a
peer like user input drawings.  This was caused by drawings being
added to the drawing history on the way into the drawing queue, but
clearing the history occurring when that drawing was pulled from the
queue.
This was partly obscurred by clearing and undoing happening in
multiple places.  More of that logic has been pulled together now in
PaintState.

The PaintCanvas has been rewritten to be a canvas, instead of managing
one.  The extra layer was unneccesary.